add active verification to login process
Adds new function isActive() to userManagement class. This will check if user is set to active (true) or not (false).

This has been added to the login logic, meaning active users can now continue to there respective dashboards (to be implemented) and none active will be notified to visit library.

New member accounts will be set to false by default requiring them to come in and be verified by staff.