add AdminVeiws and list displays
This update includes the AminView.ui, which uses the QTabWidget to allow a naviagtion between the different dashboards. They have be roughly laid out.

Where appropriate a QListWidget has been used as this allows for us to dynamically load data form JSON file in list form. This can be seen in use on manageMembers tab.