correct registration view and update displays to handle new account numbers
RegistrationView was failing to submit details as it was attempting to access nullptr causing a crash, put checks in place to make sure it is getting the correct parent when attempting to call functions.

New account generator created by @ch19517  stored values as int not string, updated the display functions to use this correctly.