add image functions
This update will include additonal functions to handle image management and display.

addBook now has logic to the handle drag and drop of a image file into a designated area. This file will then be saved as under the isbn number and stored in the /images/cover folder.

memberView ui catalogueList has been updated to dynamically create book entries based on the bookListView ui file. This is in the draft stages at this time as ratio not appearing well but it is working.