extend new user method
Have extended the adminView to be able to add new members from the appropriate dahsboard.

Created new bool userNameExits in usermanagemnt. This will check if the newusers username is already int the database and return true if so, false if not.
This has then been implemented during user registration to prevent multiple instacnes of the same username.