expand admin addMember methods to allow for creating admin/active accounts
addMemberView.ui updated with Qcheckboxs adminCheckBox and activeCheckBox.

This allows admin user to set user type and acitivity during creation step.

AddMemberView::addMemberButtonClicked has been updated to check if checkboxs are checked and set the appropriate flags to true if they are.

UserManagement::addUser() has been updated to check if QJsonObject contains the isAdmin/isActive  flags and set to them to default flase if not.

Additonal notes : cleaned and updated code where appropriate