add complete checkout, hold and returns loop for memebers.
Code has been updated to now handle the loan cycle of a library book including automated updating of "next in queue" for hold requests.

rather than store part of the hold information on a user and part on the books entry, they now store only the holdId and a seperate array has been added to the Json file to store the bulk of the information.