completed ViewBookInfo for members
When logged in as a member and clicking on a book item, the will now direct you to the book info page to see additional information such as description. This will also reflect the current status and provide options in relation to the users similar to catalogue list.

refactored code used to set availbilty and options into new function under ManagementTransaction to allow for reuse.